### 2019-12-03 #1928 @leouieda
Hi @danielskatz thank you for editing the submission. From the list of reviewers, here are a few that might be interested: @krother @dghoshal-lbl @mih Others that come to mind are the main developers of Intake, which has some similar functionality to Pooch: @martindurant and @jsignell. There is also @hmaarrfk who is implementing Pooch as the download manager for scikit-image.
