### 2019-12-04 #1928 @martindurant
The authors of Pooch ought to know about Intake ( https://github.com/fatiando/rockhound/issues/43 ), and indeed the ability of Pooch to be hooked into other systems is an important point. Will try to find time for a proper review, but don't want to commit.
