### 2019-12-04 #1928 @danielskatz
👋 @jsignell - would you be willing to review this submission for JOSS? (See https://joss.readthedocs.io/en/latest/reviewer_guidelines.html for what reviewing for JOSS)
