### 2019-12-08 #1928 @hmaarrfk
Is there a version that can be downloaded with double spacing so we may write our notes on it?
