### 2019-12-08 #1928 @danielskatz
No, sorry

Note that the [paper](https://github.com/openjournals/joss-papers/blob/joss.01928/joss.01928/10.21105.joss.01928.pdf) is fairly short, and the review is of both the paper and software, using a checklist for the [reviewer criteria](https://joss.readthedocs.io/en/latest/review_criteria.html)

JOSS reviewers generally write note either here as comments, or in the source repository as issues, or potentially pull requests on the .md file in the source repository
