### 2019-12-08 #1928 @hmaarrfk
got it, i typically go through my notes on paper and organize them again a second time, just being visual, it sometimes helps to be able to scribble things down.
