### 2019-12-08 fatiando/pooch#131 @hmaarrfk
JOSS Paper Review
xref: https://github.com/openjournals/joss-reviews/issues/1943

Looks good.

* Scientficit software is also used to "Capture" or "Acquire" data
* "The usual" -> "One common approach" when talking about smaller datasets.
* The sentence that describes the challenge with including datasets in the github repo should come before the "Larger datasets [...]" sentence.
* HTTP -> HTTPS
* You state that people are recreating the same code, but maybe give a few examples?

Paragraph 2:
* [ ] It isn't clear what the registery holds: local file name, hash, remote file url should be clearly listed.
* [ ] I "minimal dependencies" is the selling point to packagers. The selling point to users is that it is easily installable with standard python distributinos PyPi and Conda, and a wide array of python versions (2.7, 3.5->3.8).

Paragraph 3:
* [ ] the functionality to unzip files is understated and should be made more explicit. It probably deserves a paragraph of its own.

Paragraph 4:
* [ ] You should elaborate on what exactly makes intake harder to setup. It seems like Pooch is "files only", while intake is "files + metadata" which is a larger hurdle to overcome.

Paragraph 5:
* [ ] I think you can show off a little bit explaining the motivation in terms of install size this would have on scikit-image. The numbers are listed in the PR

I guess I'll go through the checklist when the review officially starts.
