### 2019-12-11 #1943 @whedon
[REVIEW]: Pooch: A friend to fetch your data files
**Submitting author:** @leouieda (<a href="http://orcid.org/0000-0001-6123-9515">Uieda, L</a>)
**Repository:** <a href="https://github.com/fatiando/pooch" target ="_blank">https://github.com/fatiando/pooch</a>
**Version:** v0.7.1
