### 2019-12-11 #1943 @danielskatz
👋 @hmaarrfk & @martindurant - here we are in the review thread

As I think you know you, your job now is go through the paper and repository and check off boxes in your checklist

Also please be sure to read the prior comments in this issue
