### 2019-12-11 #1943 @martindurant
What is the usual timeframe for this process?
