### 2019-12-12 5bf2938
Adding logging functionality to replace warnings
Add a logging object to utilities and record fetching new files
