### 2019-12-12 9712c90
Using StringIO instead of NamedTemporaryFile
