### 2020-01-02 #1943 @martindurant
Note that the links are not clickable in the github-rendered version of the PDF, only in the download. This is doubtless not a fault of the authors.

I find that the paper does indeed convey that Pooch exists, is and says how and when it might be used. It completes the purpose, therefore, of submitting to JOSS. I am still disappointed by the comparison with the rest of the field and Intake in particular. Obviously, again, I am biased here and perhaps some of the points that follow have more to do with Intake's documentation than with Pooch's, and I do not insist on any changes. However:
- the library fsspec contains local-file-cache implementations that could be seen as implementing what Pooch does, also by changing only a single line of code. I would argue that it is indeed more flexible, having many file-system backends and allowing for partial downloads of large files. It can also check the source for changes on backends that support checksum or at least write-date metadata, or general cache expiry.
- Intake does *not* require you to rewrite any code at all in the simplest case such as the pandas example, but only change one line (pd.read_csv -> intake.open_csv; same arguments) with or without local caching
- Intake does many other things too, such as read from sources which are not files, from remote servers and services; metadata and visualisation and data-source browsing. In this, I agree that Pooch does one job and does it well.
- Indeed, Intake and Pooch should have integration points, and I think this is the biggest and saddest missing point. The Pooch registry (which files) and an Intake catalog (how to load which files) clearly have overlap. Pooch could be one way of obtaining files for Intake.
- At the very least, Pooch may want to investigate fsspec and its backends, to enable downloading from a variety of file stores, not just HTTP/FTP. https://github.com/intake/filesystem_spec/blob/master/fsspec/registry.py#L12
