### 2020-01-02 #1943 @hmaarrfk
Truthfully, if pinged, I will accept for JOSS, but I think a few extra words could go a long way.
