### 2020-01-02 #1943 @danielskatz
> Note that the links are not clickable in the github-rendered version of the PDF, only in the download. This is doubtless not a fault of the authors.

Yes, this is a function of GitHub - and some of the links in the paper will not fully work until the paper is closer to being accepted, such as the link to the archived software.
