### 2020-01-02 #1943 @danielskatz
Thanks @martindurant & @hmaarrfk!

It looks to me like a bit more work would really help this paper, particularly in terms of the state of the field part.

Also, I notice both of you have not checked off the functionality box - Can you let us know what you think is still needed to enable you to check this off?
