### 2020-01-02 #1943 @hmaarrfk
I would like a checkbox for `Functionality` to be incldued in the `Software Paper` section. My comments mostly refer to that.
