### 2020-01-02 #1943 @danielskatz
I'm confused by your comment.

The software paper section includes

```Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?```

This is intended to ask if the paper sufficiently describes the functionality of the software at a high-level.

The functionality section includes

```Functionality: Have the functional claims of the software been confirmed?```

This is intended to ask if the software does what it says it does.

Can you explain what you would like again?
