### 2020-01-02 #1943 @hmaarrfk
I guess the question is:

Do you want a deeper discussion of the functionality in the paper for a more technical audience as well?

If so, some of the comments I made refer to that.
