### 2020-01-02 #1943 @danielskatz
JOSS papers are not intended to be long, and they are really intended to support/introduce the software itself and its documentation, so the paper should be more of a high-level introduction to what the software does, and the software documentation should include the deeper discussion
