### 2020-01-06 #1943 @danielskatz
@hmaarrfk - To continue our discussion, if the functionality of the software is as claimed, please check off the functionality box. If you have further thoughts on the software paper or documentation, let's talk about them as well.
