### 2020-01-06 #1943 @danielskatz
👋 @leouieda - I think we are also waiting for you to address some of @martindurant's comments on the state of the field description in the paper. Please let us know how this is going.
