### 2020-01-06 #1943 @martindurant
(just to reiterate: my comments were friendly suggestions only, and if the authors decide not to change anything, they are welcome to say so)
