### 2020-01-06 #1943 @leouieda
:wave: @danielskatz just catching up with notifications from the holidays and will start working on replies/fixes.

@martindurant @hmaarrfk thank you for the reviews! They are all appreciated and I'll provide detailed replies soon.

> (just to reiterate: my comments were friendly suggestions only, and if the authors decide not to change anything, they are welcome to say so)

@martindurant your comments are more than welcome and that is precisely why I suggested you as a reviewer :slightly_smiling_face:
