### 2020-01-06 1979924
Address comments in paragraphs 1 and 2
