### 2020-01-06 d243f97
Add reference to projects with download code
sklearn, cartopy and pyvista all have this.
