+++ paper/paper.bib
@@ -76,3 +76,25 @@
   publisher = {Zenodo},
   year = {2019}
 }
+
+@manual{cartopy,
+  author = {{Met Office}},
+  title = {Cartopy: a cartographic python library with a Matplotlib interface},
+  year = {2010 - 2015},
+  address = {Exeter, Devon },
+  url = {http://scitools.org.uk/cartopy}
+}
+
+@article{pyvista,
+  doi = {10.21105/joss.01450},
+  url = {https://doi.org/10.21105/joss.01450},
+  year = {2019},
+  month = {may},
+  publisher = {The Open Journal},
+  volume = {4},
+  number = {37},
+  pages = {1450},
+  author = {C. Bane Sullivan and Alexander Kaszynski},
+  title = {{PyVista}: 3D plotting and mesh analysis through a streamlined interface for the Visualization Toolkit ({VTK})},
+  journal = {Journal of Open Source Software}
+}
