+++ paper/paper.bib
@@ -79,7 +79,7 @@
 
 @manual{cartopy,
   author = {{Met Office}},
-  title = {Cartopy: a cartographic python library with a Matplotlib interface},
+  title = {Cartopy: a cartographic Python library with a Matplotlib interface},
   year = {2010 - 2015},
   address = {Exeter, Devon },
   url = {http://scitools.org.uk/cartopy}
