### 2020-01-07 077d0fc
Remove trailing white space
Co-Authored-By: Hugo van Kemenade <hugovk@users.noreply.github.com>
