+++ paper/paper.bib
@@ -81,7 +81,7 @@
   author = {{Met Office}},
   title = {Cartopy: a cartographic Python library with a Matplotlib interface},
   year = {2010 - 2015},
-  address = {Exeter, Devon },
+  address = {Exeter, Devon},
   url = {http://scitools.org.uk/cartopy}
 }
 
