### 2020-01-07 4dea304
https for cartopy link
Co-Authored-By: Hugo van Kemenade <hugovk@users.noreply.github.com>
