+++ paper/paper.bib
@@ -82,7 +82,7 @@
   title = {Cartopy: a cartographic Python library with a Matplotlib interface},
   year = {2010 - 2015},
   address = {Exeter, Devon},
-  url = {http://scitools.org.uk/cartopy}
+  url = {https://scitools.org.uk/cartopy}
 }
 
 @article{pyvista,
