### 2020-01-07 20d4b85
Hyphenate SHA-256
Co-Authored-By: Hugo van Kemenade <hugovk@users.noreply.github.com>
