### 2020-01-07 618ddb4
Add sentence about reducing the skimage size
