### 2020-01-07 #1943 @leouieda
> the library fsspec contains local-file-cache implementations that could be seen as implementing what Pooch does, also by changing only a single line of code. I would argue that it is indeed more flexible, having many file-system backends and allowing for partial downloads of large files. It can also check the source for changes on backends that support checksum or at least write-date metadata, or general cache expiry.

@martindurant I had seen fsspec while investigating Intake but to be honest, I didn't understand that this is what it does after reading the README or the documentation front page. In hind sight, I should have dug deeper as using fsspec would have simplified some of the work we did on Pooch.

You're right that fsspec is a much more flexible and comprehensive tool and people should definitely use it if they need this flexibility. What Pooch offers is a more limited scope and because of that it's probably easier to understand for non-specialists.

I'll include a section about fsspec in the paper as well.

> Intake does *not* require you to rewrite any code at all in the simplest case such as the pandas example, but only change one line (pd.read_csv -> intake.open_csv; same arguments) with or without local caching

Right, that is what we meant but didn't do a good job in conveying it in the text. Sorry about that. I'll revise this part to make it clear that you only need this for non-standard formats.

> Intake does many other things too, such as read from sources which are not files, from remote servers and services; metadata and visualisation and data-source browsing. In this, I agree that Pooch does one job and does it well.

I'll make sure to mention more of this in the text. We should probably also include something like this in the documentation pointing people to Intake if they need these other features.

> Indeed, Intake and Pooch should have integration points, and I think this is the biggest and saddest missing point. The Pooch registry (which files) and an Intake catalog (how to load which files) clearly have overlap. Pooch could be one way of obtaining files for Intake.

I completely agree. This has come down to a lack of time on my part to invest in really learning Intake. I'm happy to keep the door open for collaboration but given the time constraints it can't be a priority of mine (but that doesn't mean someone else couldn't take this on).

> At the very least, Pooch may want to investigate fsspec and its backends, to enable downloading from a variety of file stores, not just HTTP/FTP. https://github.com/intake/filesystem_spec/blob/master/fsspec/registry.py#L12

Right now, Pooch is pretty much a finished product since it solves the problem we initially set out to solve. But it would make a lot of sense to replace our download code with fsspec in the future and use it as our backend. That way, we can contribute back new backends to fsspec instead.
