### 2020-01-08 7d50cb4
Update paper/paper.md
Co-Authored-By: Mark Harfouche <mark.harfouche@gmail.com>
