### 2020-01-09 11379de
Allow to use different hashing algorithms
