### 2020-01-10 c0a044e
Edits to JOSS paper after review by Mark Harfouche (#132)
Changes to the paper only to address comments by the reviewer. 
Mostly clarifying the advantages of Pooch.

Fixes #131
