### 2020-01-10 1e2ea0e
Style fixes
Co-Authored-By: Leonardo Uieda <leouieda@gmail.com>
