### 2020-01-10 820fab6
Revert file_hash to original API signature
