### 2020-01-10 fe1336b
Add docstring for add_hash_algs
