### 2020-01-15 dcc2395
Capitalize NumPy
Co-Authored-By: Hugo van Kemenade <hugovk@users.noreply.github.com>
