### 2020-01-15 7616cc7
Clarify fsspec/Intake part and how Pooch compares
