### 2020-01-17 92da0fd
Expand JOSS paper section on Intake/fsspec (#134)
As pointed out by reviewer Martin Durant, we failed to mention fsspec in
the "state of the field" section and could have done a better job at
presenting Intake. Include a couple of sentences more on Intake and
fsspec, how fsspec could be used in Pooch in the future, and what are
the advantages and disadvantages of Pooch with respect to these tools.
