### 2020-01-17 #1943 @leouieda
Hopefully all reviewer comments have been addressed in fatiando/pooch#132 and fatiando/pooch#134
