### 2020-01-17 b1fb777
Fix capitalization in JOSS paper references
Some of the Bibtex was poorly formated to preserve capitalization
(Python vs python).
