### 2020-01-17 16efeca
scikit-image is lower case
