+++ paper/paper.bib
@@ -11,7 +11,7 @@
 }
 
 @article{scikit-image,
-  title={scikit-image: image processing in {Python}},
+  title={{scikit-image}: {image} processing in {Python}},
   author={Van der Walt, Stefan and Sch{\"o}nberger, Johannes L and
           Nunez-Iglesias, Juan and Boulogne, Fran{\c{c}}ois and Warner,
           Joshua D and Yager, Neil and Gouillart, Emmanuelle and Yu, Tony},
