### 2020-01-17 7d89556
Fix capitalization in JOSS paper references (#135)
Some of the Bibtex was poorly formated to preserve capitalization
(Python vs python).
