### 2020-01-17 #1943 @leouieda
Should be fixed now. The bibtex was poorly formatted.
