### 2020-01-17 f6eb409
Changelog for v0.7.1
This release marks the end of the JOSS paper review.
Minor release with only error message improvements.
