+++ v0.7.1/_sources/index.rst.txt
@@ -1,0 +1,21 @@
+.. title:: Home
+.. raw:: html
+
+    <div class="banner">
+        <img src="_static/readme-banner.png" alt="Pooch">
+        <h2>A part of the <a href="https://www.fatiando.org/">Fatiando a Terra</a> project.</h2>
+    </div>
+
+.. include:: ../README.rst
+    :start-after: placeholder-for-doc-index
+
+.. toctree::
+    :maxdepth: 2
+    :hidden:
+    :caption: Documentation
+
+    install.rst
+    citing.rst
+    usage.rst
+    api/index.rst
+    changes.rst
