### 2020-01-17 #1943 @leouieda
@danielskatz  done. Release version 0.7.1 and archive DOI [10.5281/zenodo.3611376](https://doi.org/10.5281/zenodo.3611376)
