### 2020-01-17 57d2747
Update citation to JOSS paper
Add a changelog entry for another minor release to update the citation
information and page. No DOI will be issued for this release.
